Move LocationInfo out of Engine/Graphics
It holds respawn counts, reputation and the alert flag, none of which is
graphics. The header goes to Engine/ and the four includers follow. The
other TODO in the file, about renaming the type with a Map- prefix to tell
it apart from MapInfo, is a separate question and stays.
