+++ src/Engine/Graphics/Indoor.h
@@ -10,6 +10,7 @@
 #include "Engine/mm7_data.h"
 #include "Engine/EngineIocContainer.h"
 #include "Engine/SpawnPoint.h"
+#include "Engine/LocationInfo.h"
 
 #include "Core/Time/Time.h"
 
@@ -17,7 +18,6 @@
 #include "Library/Geometry/Plane.h"
 #include "Library/Geometry/BBox.h"
 
-#include "LocationInfo.h"
 #include "LocationFunctions.h"
 #include "FaceEnums.h"
