+++ src/Engine/Graphics/LocationFunctions.h
@@ -1,8 +1,7 @@
 #pragma once
 
 #include "Engine/MapEnums.h"
-
-#include "LocationInfo.h"
+#include "Engine/LocationInfo.h"
 
 // TODO(captainurist): move to Engine/ and drop the Location- prefix, should be MapSmth.
