+++ src/Engine/LocationInfo.h
@@ -1,7 +1,6 @@
 #pragma once
 
 // TODO(captainurist): we also have MapInfo, rename this into something more sane with Map- prefix.
-// TODO(captainurist): this isn't graphics, it belongs in Engine/ rather than Engine/Graphics/.
 struct LocationInfo {
     int respawnCount = 0; // Number of times a location was respawned, including the initial spawn.
     int lastRespawnDay = 0; // Day of the last respawn (days since GameTime zero to last respawn).
